refactor: inputMapper 제거
