+++ src/main/java/controller/LadderGameController.java
@@ -1,24 +1,25 @@
 package controller;
 
-import domain.*;
+import domain.Height;
+import domain.Ladder;
+import domain.Names;
+import domain.NonContinuousLineGenerator;
 import view.InputView;
 import view.ResultView;
 
 public class LadderGameController {
 
     private final InputView inputView;
-    private final InputMapper inputMapper;
     private final ResultView resultView;
 
-    public LadderGameController(InputView inputView, InputMapper inputMapper, ResultView resultView) {
+    public LadderGameController(InputView inputView, ResultView resultView) {
         this.inputView = inputView;
-        this.inputMapper = inputMapper;
         this.resultView = resultView;
     }
 
     public void run() {
-        Names names = inputMapper.mapToNames(inputView.readNames());
-        Height height = inputMapper.mapToHeight(inputView.readHeight());
+        Names names = inputView.readNames();
+        Height height = inputView.readHeight();
 
         Ladder ladder = Ladder.createFrom(new NonContinuousLineGenerator(), names.getNameCount(), height);
